fix: address Codex review round 2 (2 critical, 5 high, 1 medium)
Critical:
- eidos/models.py: reload_eidos_from() now merges with values section
  (max_steps was regressing from 40 to 25 after hot-reload)
- outcome_log/advice_feedback/implicit_outcome_tracker: atomic temp-write
  + os.replace for JSONL rotation (no more in-place truncate race)

High:
- advisory_engine.py: safety check moved pre-emit (was post-emit),
  fallback emit path now also runs safety check
- auto_tuner.py: int() → float() for quality_threshold (recurring bug)
- advisory_engine.py: _fallback_guard_allows() uses _tail_jsonl() instead
  of reading entire file on every tool call
- outcomes/linker.py: hash() → hashlib.sha1 for deterministic IDs
  across Python processes

Medium:
- validate_and_store.py: wire flow section reload to reset_enabled_cache()
- test_advisory_profile_sweeper.py: skip if benchmarks/ not present

+++ lib/auto_tuner.py
@@ -377,7 +377,7 @@ class AutoTuner:
 
         # --- Meta-Ralph quality threshold ---
         mr_cfg = tuneables.get("meta_ralph", {})
-        current_qt = int(mr_cfg.get("quality_threshold", 4))
+        current_qt = float(mr_cfg.get("quality_threshold", 4))
         if health.feedback_loop_closure < 0.3 and health.cognitive_growth > 50:
             recs.append(TuneRecommendation(
                 section="meta_ralph", key="quality_threshold",
